Update pom.xml, fix kafparser and offset attr issue
The pom.xml has been updated to get the current opennlp-tools lib.
The opennlp-tools-1.5.3 is no longer a SNAPSHOT.
By other side, the kafparser in use is the Kyoto one.
There was an issue with the offset attr inside worforms.
Anyway, the kyoto parser should be replaced by a better one.